COMPMID-417: Fix assert in GEMMTranspose
The assert was checking the wrong thing. Only if the window over the
input is smaller than the number of processed elements, the output shape
would be empty.

However, the valid region will be empty if the input's first dimension
is less than the number of elements processed. That required the changes
in TensorShape.

Change-Id: I36fed7893dfd502e26c5c776c9a2d774d6cd91c6
Reviewed-on: http://mpd-gerrit.cambridge.arm.com/79813
Tested-by: Kaizen <jeremy.johnson+kaizengerrit@arm.com>
Reviewed-by: Gian Marco Iodice <gianmarco.iodice@arm.com>