COMPMID-745: Handle bilinear interpolation for remap differently
Until now if pixels coordinates were clamped to the padding, then the
bilinear interpolation would be performed using using 4 pixels out of
bounds, requiring 2 pixels in the right and bottom borders.

Instead now we skip the interpolation if it's not needed.

When using bilinear filtering + border mode constant: we need two pixels on the right and bottom for the interpolation

- Reverted to BorderSize(1)
- Handle special case differently in bilinear function

Change-Id: Ic431975cff236c048fa329644ae7b67d4476ab1e
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/113016
Reviewed-by: Georgios Pinitas <georgios.pinitas@arm.com>
Reviewed-by: Michalis Spyrou <michalis.spyrou@arm.com>
Tested-by: BSG Visual Compute Jenkins server to access repositories on http://mpd-gerrit.cambridge.arm.com <bsgcomp@arm.com>