COMPMID-1410 (Nightly) Mismatches in CLMeanStd function for float
Change CLReductionOperation border to be multiple of 64 instead of 16.
The opencl kernel works only with local_size(0) being a power of 2. This will
generate a padding of 63 if input_width % 64 = 1, but I don't think it's a
big issue and it keeps the border calculation pretty simple.

Also, increased tolerance for fp32 because there were mismatches
for the 4K image.

Change-Id: Id44990a262b2d6eff4c8ce56eb7c886274d9847e
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/143415
Reviewed-by: Pablo Tello <pablo.tello@arm.com>
Reviewed-by: Anthony Barbier <anthony.barbier@arm.com>
Tested-by: Jenkins <bsgcomp@arm.com>