[MLCE-54] graph_alexnet example throw an exception
The 32bit version of AlexNet seems to require ~1.4 GB of memory to run (got this
result via valgrind's massif tool).

The problem is a miscalculation of the shape of the _transformed_b tensor, which
should not be multiplied by the size of the data type, but rather represent only
the size of the dimensions.

Change-Id: I55dfb538b31772afc4ce0ce0ccd31fdeb1585cfb
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/147666
Reviewed-by: Georgios Pinitas <georgios.pinitas@arm.com>
Reviewed-by: Anthony Barbier <anthony.barbier@arm.com>
Tested-by: Jenkins <bsgcomp@arm.com>