Use .inst directives instead of .int directives.
Has a couple of benefits, one is a disassembler that actually
understands dot product will start showing the dot product
instruction for what it is rather than just a random .word.
For the interested parties in actually why compilers and toolchains
manage to disassemble this , please go and look up mapping symbols
from toolchains.

Secondly .word is a data directive and if you ever have a customer
run Arm compute library on big endian, on AArch64 this will not work.
This is because data on big endian is well, big endian but the code
section is not big endian but just little endian. Admittedly there
will be many other things that need to be fixed for big endian
to work reliably.

Eyeballed satisfactorily with a simple case. If someone
could run this through a test run with the CI that would be
great.

Thanks,
Ramana

Change-Id: I0b9573ecbed298afc967d675b0542a6fe72b4c52
Signed-off-by: Ramana Radhakrishnan <ramana.radhakrishnan@arm.com>
Reviewed-on: https://review.mlplatform.org/c/1588
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Georgios Pinitas <georgios.pinitas@arm.com>