COMPMID-2204:  RuntimeContext interface for NEON functions.
This patch creates the interfaces for the runtime context for NEON.

Only the Neon backend implements the context which currently only holds
an instance of the scheduler.

The NEActivationLayer function has been updated to use the new context
interface and the corresponding validation tests ported.

Change-Id: I32e7e6aa888796dcbbfc5039b1e7f784a24f47da
Signed-off-by: Pablo Tello <pablo.tello@arm.com>
Reviewed-on: https://review.mlplatform.org/c/1851
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Michele Di Giorgio <michele.digiorgio@arm.com>