COMPMID-2563 Change how the best overall GEMM configuration is selected
* Based on a specified tolerance, each GEMMParam (GEMM Shape) now can
have a set of best GEMM configurations, instead of just a single one.
This improves the robustness and completeness of the tuned results, and
is in compliance with how we define the GEMMParam archetypes (the main
goal of this story)

* The tuner then tries to find the best overall GEMMConfig, from all the
best config sets, through the same voting mechanism: the config that
gets voted the most is the best overall GEMMConfig.

Change-Id: Ief770bb6ffc04629d91f1dc778eea69274e007f0
Signed-off-by: SiCong Li <sicong.li@arm.com>
Reviewed-on: https://review.mlplatform.org/c/2228
Reviewed-by: Georgios Pinitas <georgios.pinitas@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Gian Marco Iodice <gianmarco.iodice@arm.com>