COMPMID-2939 Keep benchmark_examples and validate_examples
Remove FIXME comments in Sconscripts referring to the removal of
benchmark_examples and validate_examples, as we are keeping them in
releases from now on.

This is a 19.11 release activity ticket but the patch needs to be
backported to master because this applies to all future releases.

Signed-off-by: SiCong Li <sicong.li@arm.com>
Change-Id: Ifac73ebeaf12c93d6309f73bb011bd275e1fe503
Reviewed-on: https://review.mlplatform.org/c/2334
Reviewed-by: Michele Di Giorgio <michele.digiorgio@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>