COMPMID-3324: Allow in-place mutation on nodes with separate multiple inputs
The current implementation of the in-place mutator is limited by the
assumption that each node only produces one output and therefore all
its output edges represent the same tensor. This means that doing
in-place computations on it, would modify the input of other nodes.

However, this is not the case for operators like split, which creates N
different tensors, one for each output edge.

This patches relaxes this limitation by checking whether the tensor
related to a specific output edge is the same as any of the other edges.
If this is the case, then in-place computation cannot be done.

Change-Id: I3302b41b1f6ec3b4f2ac425ba381778f1c0a4f31
Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/3421
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Manuel Bottini <manuel.bottini@arm.com>
Reviewed-by: Georgios Pinitas <georgios.pinitas@arm.com>