COMPMID-3774: Investigate NEReduceMean QASYMM8_SIGNED nightly failure
Tolerance issue due to requantization. The NEON implementation does all
computations in float when input and output quantization info are
different and reduction on multiple axes is required. On the other hand,
the reference performs the first reduction in float, then requantizes
and then performs the remaining reductions in the quantized domain using
the output from the first redcution. This causes small discrepancies in
few cases, hence increasing the tolerance.

Change-Id: Ib862f599ce3924cbad651bab77227d52e15eff88
Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/3937
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Georgios Pinitas <georgios.pinitas@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>