COMPMID-3773: Fix the accumulation of projection in QLSTM
The "output_state_in" (previous output state) tensor
is used for accumulation of projection.

The argument for the tensor given to configure() has
to be changed to non-const since CLTensor needs
to be non-const for map() function call for data copying.

Even though NEON-side doesn't need the same change,
it has been done for consistency.

Change-Id: Ifba0ab6dc8260c468e9f087bf51824daefbab7a3
Signed-off-by: Sang-Hoon Park <sang-hoon.park@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/4018
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Georgios Pinitas <georgios.pinitas@arm.com>