Add tolerance for quantized activations computed in float
Some of the activation functions need complex mathematical operations
and are implemented by dequantizing to float, performing the activation
in the float domain and requantizing back. In such cases, the results
may differ slightly between reference and optimized code.

In fact, when running validation through valgrind we get a difference of
1 in the results and therefore, an absolute tolerance of 1 is added to
the tests.

Resolves: COMPMID-4067

Change-Id: Ic2eca5616371b0a324a246d40b515ddc9f576e61
Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/4841
Reviewed-by: Giorgio Arena <giorgio.arena@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>