Fix NeDepthwiseConvolution bad_alloc issue
The memory allocation of the failed tests's tensors are huge, all the virtual memories are used up. And then when assign the output to _reference, a copy constructor is called and allocate more memory and caused the crash.

Delete one of the tensor shape and adjust one.

Resolve: COMPMID-4366

Signed-off-by: Sheri Zhang <sheri.zhang@arm.com>
Change-Id: I91a816384ec358481018a724a7b0cd59730698fc
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/5426
Reviewed-by: Michele Di Giorgio <michele.digiorgio@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>