Allow clearer suppression of toolchain and compiler prefixes
Currently it is counterintuitive to disable toolchain prefixes: we need
to pass an empty space to it; passing an empty string would not disable
prefixes but instead instruct the build script to use a set of default
prefixes.

With this patch we restore the intuitive approach of passing an empty
string "" to disable the prefix.

Resolves COMPMID-5353, COMPMID-5380

Signed-off-by: SiCong Li <sicong.li@arm.com>
Change-Id: I2fe84689df9093cf6baf507dde44ca5ebf61023b
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/7686
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Gian Marco Iodice <gianmarco.iodice@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>