Rework heuristic in ClConv2d
The heuristic has been tweaked to call direct convolution when we think
it can be faster than gemm-based convolution. The main change is
affecting the selection of the convolution method on the first layer.

In general, the question we should ask for the first convolution layer of
a model is: when the execution time of im2col + gemm < direct?. Since
im2col does not depend on the OFM, it means that when OFM is big enough,
the contribution of im2col is small and the GEMM approach is preferable.
From internal experiments, the OFM threshold is 64.

Resolves COMPMID-5504, COMPMID-5504, COMPMID-5477

Change-Id: If1bd1fa93c185ffa874388e29866244e62ca3494
Signed-off-by: Gian Marco Iodice <gianmarco.iodice@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/8231
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Gunes Bayir <gunes.bayir@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>