Fix unsupported configuration in CLFullyConnected validation
When the weights to CLFullyConnected layer are not constant and the weights need reshaping, we prefer MatMul kernels instead of Gemm-based ones. The bias addition is currently unsupported in MatMull, but the validate() function does not account for this properly. This patch fixes the validation of this function.

Resolves: COMPMID-6338
Change-Id: I5c240191ae8e369753691c43ab4a30d4ae1776b0
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/9882
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Viet-Hoa Do <viet-hoa.do@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>