Fix out-of-scope CLBufferMemoryRegion's buffer still in queue issue
When a CLBufferMemoryRegion is freed, it also frees its cl::Buffer
object. At this point we need to flush the queue to ensure all prior
commands that may use this buffer are completed before the buffer's
deallocation.

Previously a CommandQueue object is owned as a member inside
CLBufferMemoryRegion. Whenever CLBufferMemoryRegion is freed it causes
the queue to be released, which implicitly flushes the queue.

Now we need to explicitly flush the queue, without the excessive
releasing of the queue

Resolves COMPMID-6492

Signed-off-by: SiCong Li <sicong.li@arm.com>
Change-Id: I799507bcff8526d1381cde53d7c6298684c6d3ee
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/10126
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Gian Marco Iodice <gianmarco.iodice@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>