arm_gemm: convolution: optimize convolver.hpp.
The code in convolver.hpp generates pointers into either the
appropriate point in the input activation tensor or the padding buffer
for each kernel point of each output point of the convolution.  This is
done at runtime interspersed with the data transform and matrix
multiplication steps.  As such, it can have a significant impact on
performance, particularly for low input channel counts.

This change improves the performance of this code by streamlining the
checks for out of range input points (which must be directed to the
padding buffer).  The previous implementation checked all four borders
for every point.  The revised code does the checks one at a time, and
for any failing check applies the result to as many output points as
possible without repeating the other checks.

Signed-off-by: David Mansell <David.Mansell@arm.com>
Change-Id: I36a4fa114b425c1bcba2be40acf36718522519f5
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/11004
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Gunes Bayir <gunes.bayir@arm.com>